requirements.txt
firecrawl-py
azure-cognitiveservices-speech
openai